requirements.txt
virustotal-api==1.1.11
redis==3.5.3
retrying==1.3.3
flask==1.1.2
